Introducing issues templates for bugs and feature requests
When creating bug reports inside this repository would be nice to have a
template to follow. This will save developers time when debugging the
problem, since most of our requests don't have a reproducible today. The
same for new features requests, since some features are small and don't
need a entire blueprint.

Signed-off-by: Beraldo Leal <bleal@redhat.com>