utils.message: Fixes avocado log level
Currently we are only logging a few messages from utils because the
logger level is not set properly. Let's set the log level so those
messages will started to be populated.

Also, by default propagate is set to True, which will make events logged
into this logger be passed to the handlers of higher level (ancestors)
loggers. Setting this to false, logging messages are not passed to the
handlers of ancestor loggers. This is fixing the duplicate messages that
appeared after the first fix mentioned above.

Signed-off-by: Jan Richter <jarichte@redhat.com>
Signed-off-by: Beraldo Leal <bleal@redhat.com>