Safeloader: filter out properties from test methods results
Avocado can not currently run properties as tests, so it does not make
sense to present those as valid test methods.

Althought this change makes sense in itself, it is also related to the
existence of the "teststmpdir" property in the "avocado.Test" class,
and the fact that it should not be visible as a test method.

In the future, support for using properties as tests may be added,
along with the removal of "teststmpdir" as stated in eb5190962.

Signed-off-by: Cleber Rosa <crosa@redhat.com>