Python unittest support: use SKIP status for skipped tests
Python's unittests do not have the concept of canceled tests, only
skipped tests.  But, Avocado has been using "CANCEL" as a status for
Python unittests, possibly out of the reasoning that Avocado itself
ran "something" in its "avocado.core.test.PythonUnittest" class to get
to the point of determining the unittest status.

Looking closely, and comparing to the nrunner behavior, I believe the
current approach is wrong.  A Python unittest that is skipped, should
be reported as SKIP, and not as CANCEL.

Signed-off-by: Cleber Rosa <crosa@redhat.com>