Selftests: print test logs for failures, errors and interruptions
The selftests are run via "check.py" on a number of different
environments, including ones which are hard or even impossible to
debug interactively.

To ease debugging, or sometimes enable any debugging whatsoever, lets
always print the test logs on any "not OK" situation.

Also, include standalone and "raw" stdout and stderr files.

Signed-off-by: Cleber Rosa <crosa@redhat.com>