Allow customization of the Python command used in "make develop"
Because of perceived and verified limitations on setuptools, which
installs dependencies of packages, but does not make them fully
available in the installation environment, let's allow users to
customize the command to be used.

This allows us to force a "pip install" based installation on
environments such as readthedocs.org, which may have conflicting
packages.

Fixes: https://github.com/avocado-framework/avocado/issues/5027
Signed-off-by: Cleber Rosa <crosa@redhat.com>