Safeloader: do not keep checking directives on the hierarchy
Because the Avocado safeloader starts with the class definitions on
the file asked to be parsed, and walks its way up on the hierarchy,
it needs to take favor the status of the class definition itself to
control the overall behavior.

This is true and common for tests that define base classes, and then
specialized classes.  The base classes, are usually protected with
an ":avocado: disable" docstring directive.  If, we keep looking
at the base class dicstring directive when walking up the class
hierarchy of a specialized class, we can find an ":avocado: disable"
docstring directive that is not intended for the specialized class.

This changes the check that is performed after we leave the primary
class definition, ignoring the docstring directives after that point.

Fixes: https://github.com/avocado-framework/avocado/issues/5001
Signed-off-by: Cleber Rosa <crosa@redhat.com>