Resolver: allow avocado.core.Test tests to be found as python-unittest
Since 48b34c881, properties such as "teststmpdir" from avocado.Test
are never found on avocado-instrumented nor on python-unittest.  This
allows us to remove those properties from the list of methods that
should be found as python-unittest test methods.

With regards to the other tests, they should indeed continue to be
found as *python-unittest*s but not as avocado-instrumented.  The
reason is that these tests inherit from "avocado.core.test.Test",
which happens to be a private location of the "avocado.Test"
implementation.  For the safeloader, it simply means it's *not*
an avocado-instrumented tests.

Likewise, because the class "avocado.core.test.Test" inherits from
"unittest.TestCase", it should be found by the *python-unittest*
resolver plugin.

In short, the changes in eb5190962 were conceptually broken and did
not account for the multiple plugins resolving (or not) tests from the
same file.  This also adds more functional resolver tests to ensure
that the avocado-instrumented and python-unittest individual behavior
are correct.

Signed-off-by: Cleber Rosa <crosa@redhat.com>