Docs: add --loader whenever result is for the loader
Most of the command examples and output default to ones produced by
the loader, but some are missing the explicit command line.

Because at this stage, it'd be too invasive to change all the mentions
of "INSTRUMENTED", or "SIMPLE" (and others), let's add the command
line option so that command and output match.

Signed-off-by: Cleber Rosa <crosa@redhat.com>