spawners: Improve spawner's API with new class
Current Spawners API does not support the separation between the ones
that need deployment and the ones that don't. This introduce one new
class: DeploymentSpawner. It will be sed to deploy the bootstrap
artifacts (Avocado runners, tests, and test's data). i.e: PodmanSpawner
and any other remote spawner like a future SSHSpawner. This is a
prep-work providing a better organization giving basic requirements for
future implementations.

Related to #4932.

Signed-off-by: Beraldo Leal <bleal@redhat.com>