[nrunner] add access for EXEC test to some AVOCADO_* env variables
This adds support to the AVOCADO_VERSION and AVOCADO_TEST_WORKDIR to the
EXEC test on nrunner.

Other variables, like AVOCADO_TEST_BASEDIR, AVOCADO_TEST_LOGDIR and
AVOCADO_TEST_LOGFILE may not make sense as they would hold the
information where the avocado command started and not the environment
where the test is currently running.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>