exec-test: allow a fixed command to be defined
And with that, allow the use of references as the positional argument
to the defined command.

This allow users to achieve the same behavior as the "external runner"
feature of the legacy runner.

With the removal of the legacy runner, we can evaluate if a layer of
compatibility for the "--external-runner" command line option should
be added on top of this feature.

Signed-off-by: Cleber Rosa <crosa@redhat.com>