Fix auto status servers with long dirs
Using the job's result directory (AKA logdir) for the status server
has advantages because a job will always have that directory.  But it
can cause issues with paths that are "too long" for UNIX domain
sockets.

This changes the location of the socket for automatically configured
status servers, in an attempt to never get close or exceed the path
size mentioned above.

Fixes: https://github.com/avocado-framework/avocado/issues/5062
Signed-off-by: Cleber Rosa <crosa@redhat.com>