Python set type available for the job config record
For saving the records of job config, we are using the json decoder,
which unfortunately doesn't support the python `set` type. Because of
that problem, every `set` in job configuration will be lost in the
record. Right now we are using a `set` for `core.show` attribute and
this attribute is not saved in job config records which causing troubles
when we are trying to recover config from the record. As an example is
`replay` command which is using old configuration based on it's record.
This commits reserve a dict with key `__set__` as a flag for `set` type
inside job configuration record. With this change the `sets` can be
saved and restored from job configuration records.

Signed-off-by: Jan Richter <jarichte@redhat.com>