Job results data: have a separate variants for each test suite
The "jobdata/variants.json" file is currently saving a list of
variants, one for each test suite.  But, this is not the format
that is expected from it, as it can be seen with any attempt to
use it via "--json-variants-load".

This will create a single variants JSON file, named sequentially, and
if available, with the name of the suite too.

We could attempt to preserve compatibility with the "variants.json" by
creating a link with the first suite's variants, but, given that this
has been broken for a while, I don't expect many users to be relying
on it.

Also, further work to suite specific information preserved separately,
such as the suite's configuration (versus the global job configuration)
should be discussed and evaluated.

Fixes: https://github.com/avocado-framework/avocado/issues/5068
Signed-off-by: Cleber Rosa <crosa@redhat.com>