selftests/check.py: set shorter and more relevant variant IDs
The Job API tests using dict_variants have many values that are not
relevant for the unique identification of each test.  Some of them are
just data that the test needs.

This chooses a better set of values to use in the name of variants,
allow for much more comprehensible test results.

Signed-off-by: Cleber Rosa <crosa@redhat.com>