utils.vmimage: fix default version for Ubuntu
The current default version of Ubuntu is 21.10. When converting
this version to float, the last 0 is removed and this leads
to try to fetch version 21.1 instead of 21.10
Overriding get_best_version() to take care of that.

Remove overrided get_versions() since it's not longer required.

References: https://github.com/avocado-framework/avocado/issues/5073

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>