Dict variants: check the variant id keys content
Using the configuration dictionary default value as a fallback, but
that is a mistake because the 'run.dict_variants.variant_id_keys' will
always exist (it's registered by the Init plugin).  So
its ('run.dict_variants.variant_id_keys') default value (an empty
list) will be returned.

There needs to be a check for the content of that configuration item,
and not really its presence.

Signed-off-by: Cleber Rosa <crosa@redhat.com>