nrunner interface: add checks and fixes for bare "runnable-run"
This is a simple interface check for what happens when one runs
"avocado-runner-something runnable-run".

For robot and golang, it triggers two crashes, which are also fixed
here.

One special characteristic of "avocado-runner" is that it is capable
of multiple runnable kinds, of one must specify the runnable it should
run.  The existence of "avocado-runner" itself is debatable, but at
this point the knowledge of requirements (from Avocado's standpoint)
is based on the "capabilities" output, and based on the script name.

Either way, this adds a more consistent behavior for python-unittest,
golang and robot, while documenting the current behavior.

This is part of #4058.

Signed-off-by: Cleber Rosa <crosa@redhat.com>