avocado-instrumented: use the complete URI as the test name
When compared to the behavior under the legacy runner, the
avocado-instrumented runner is only using the class and method name
for the Test ID, as seen by the avocado.Test class.

This means different output for important lines such as "PASS" in the
test log when comparing the two runners.

This change to use the full URI makes the avocado-instrumented runner
produce output similar to the legacy runner.

Signed-off-by: Cleber Rosa <crosa@redhat.com>