`--store-logging-stream` fix for the nrunner
The `--store-logging-stream` option wasn't available for the nrunner.
This commit is fixing this. The change from the legacy solution is that
the streams are not saved inside the job results directory, but they are
split by the tests inside test-result directory. So each test has its
own files with the user stream logs.

Reference: #5081
Signed-off-by: Jan Richter <jarichte@redhat.com>