loader/test: remove legacy runner PythonUnittest support
This is the beginning of the dismantling of the loader, one test type
at a time.  This removes the Python Unittest support, because, among
other reasons, it's built on top of the external runner feature, that
is legacy-runner specific and will also be removed.

The relevant tests have been ported to become resolver tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>