contrib: replace external-runner based script with Job API one
The "avocado-run-inplace.sh" script addresses the use case where one
wants one command to be run under Avocado, giving the command right
from the command line.

Because the external-runner, in its current incarnation at least, is
about to be removed, let's supply a rough substitute.

The actual name of the "test", as visible by the user, can be improved
when issue #4980 is resolved.

Signed-off-by: Cleber Rosa <crosa@redhat.com>