TestError: make it a more private class
The "avocado.core.test" module contains a collection of test classes,
most of which were intended to be user facing.  TestError is one that
is very private, and only used at one specific location.

With the move towards nrunner only, we don't need to keep those
classes as a way to signal results to the runner.

So, let's move TestError closer to its sole user, and eventually with
the removal of the loader, look at other ways to implement the test
error handling notification.

Signed-off-by: Cleber Rosa <crosa@redhat.com>