Port loader SIMPLE tests to resolver exec-test
This ports the loader's SIMPLE tests to the exec-test counterparts.
When the test seems to be a duplicate, or not relevant to the
resolver, it's removed.

In this particular case, there are tests that do the same for shell
based and Python based excutable tests, which IMO does not add to the
coverage and are duplicates.

Signed-off-by: Cleber Rosa <crosa@redhat.com>