selftests/check.py: set a category for the results of selftests
When developing Avocado, it's common to run a number of jobs.  But,
more often than not, we resort to "./selftests/check.py" (or "make
check").

When the results for those jobs get mixed with every other job
results.  By setting a category, one can easily go into
~/avocado/job-results/avocado-selftests/ and browse through the last
runs of selftests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>