Remove example "doublefree" tests
These tests were intended to show how one can compile a binary as part
of the test, run it and look at what it produces.  But, these tests
produce output that is very dependent on the environment.  There are
already conditionals for different OSs, and newer gcc/glibc will
require yet another conditional.

On gcc 11.2.1 and glibc 2.34, the error message produced is:

  free(): double free detected in tcache 2

IMO those provide little value as example tests, observing the actual
Avocado use cases, and are too dependent on environment, so they
better go.

Signed-off-by: Cleber Rosa <crosa@redhat.com>