Yaml To Mux: use nrunner in params test
The actual Avocado runtime variables that will be passed to an nrunner
exec-test as environment variables is still being discussed.  So, to
be able to run the parameters being seen as environment variables,
let's add a simpler example tests that can be used on the selftest.

Reference: https://github.com/avocado-framework/avocado/issues/4742
Signed-off-by: Cleber Rosa <crosa@redhat.com>