Yaml to Mux: convert dry-run and mux-inject test
The test in question, mixes two different features.  When converting
that to the nrunner, we must face the fact that a test (in this case
avocado.core.test.DryRunTest) is actually *never* started in dry-run
mode (the runnable becomes of type dry-run).

Because of that, and because they are unrelated to the Yaml To Mux
varianter, the dry-run checks are abandoned here.  In place, we add a
test that list parameters, and check for the effects of mux-inject
only.

Signed-off-by: Cleber Rosa <crosa@redhat.com>