Convert test parameters to variants in tasks.
After the 050bc330dad4cbe59eafc564b561b727d246a3a5 the suite will raise
an error when the variants and test parameters are used together. With
this in mind, there is no difference between one variant and test
parameters from the task perspective. Let's add test parameters to test
tasks as a variant. In the future, we can discuss if the `variant` is
the right variable name in tasks for holding data about parameters.

Signed-off-by: Jan Richter <jarichte@redhat.com>