Reorg makefiles, add makefile with packaging rules
Reorganize makefiles to create Makefile.pkg with all
the variables and targets related to rpm packaging.

Remove completely variables RELEASE_COMMIT and
RELEASE_SHORT_COMMIT, since they are not used.

Remove requirements-plugins target, it's not used.
None of the plugins under optional_plugins provides
a requirements.txt file or a Makefile with
a requirements target.

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>