Update distro.py
Signed-off-by: shilei shileib@uniontech.com
Add probe for uos systems in general
UnionTech is the name of the company ,uos is the name of the operating system.
UnionTech Software Technology Co., Ltd. is a domestic basic software company, which was jointly established by operating system manufacturers in 2019 [3]. The company focuses on the development and service of operating system and other basic software, and is committed to providing operating system products and solutions for users in different industries.