runnable: introduce custom format f-strings
With this experimental change, we are enabling users to set custom
identifiers for the runnables/tests. By default we are still using {uri}
as identifier but any custom f-string using uri, args and/or kwargs can
be used. For now we have a global settings for all runners, since we
have the ability to create custom suites. However future improvements
can be added here.

Fixes #4980.

Signed-off-by: Beraldo Leal <bleal@redhat.com>