spawners.podman: adapt podman spawner to use eggs
This commit is changing the Avocado deployment model for this
"DeploymentSpawner" (Podman). Currently only nrunner is being deployed
to the isolated environment. With this patch, we are downloading (and
caching) two eggs: avocado and setuptools. Those eggs are then copied to
the container before starting it with a new entry point.

This is part of our new strategy to have a better Avocado support inside
the container when running tests in isolated environments.

Fixes #4933

Signed-off-by: Beraldo Leal <bleal@redhat.com>