Registering avocado egg during install process
This is an experimental ideal to fetch and register avocado eggs during
the install process, for now it is only working when using the develop
mode, but if approved I will adapt the normal installs to do the same.

Basically this is downloading our egg for future use (i.e podman
spawner) so we can deploy avocado using local resources.

It is more likely users will have Internet connection  when installing
this software then when running it. This is an attempt to minimize
connectivity issues, so we can download only in case the asset is not
found.

This is part of #4934.

Signed-off-by: Beraldo Leal <bleal@redhat.com>