contrib.scripts: new script to fetch eggs
This is an experimental idea to fetch and register Avocado's eggs a
priori to speed up Avocado deployment process when using deployment
spawners.

Basically this script it will download our eggs for future use (i.e
podman spawner) so we can deploy avocado using local resources.

This is part of #4934.

Signed-off-by: Beraldo Leal <bleal@redhat.com>