avocado/core/teststatus.py: add a handy list of OK and NOT OK status
When checking if a test status is "good" or "not good", it's better to
refer to a canonical source for truth.  That currently exists as the
`STATUS_MAPPING` dictionary, but it's not straightforward to use.

Let's add a list of both "OK" and "NOT OK" statuses to make its usage
simpler.

Signed-off-by: Cleber Rosa <crosa@redhat.com>