Set the exec-runner environment
The 'AVOCADO_TEST_OUTPUTDIR' variable was always inside the
`runnable.kwargs` which caused that the exec-runner environment was
always set. But in the commit 080653d4871783ec051e38477210997440b9e5af we are removing the
'AVOCADO_TEST_OUTPUTDIR' from the `runnable.kwargs`, which means that
the `runnable.kwargs` might be empty and the exec-runner environment
might be `None`. This commit will set the exec-runner environment even
the `runnable.kwargs` is empty.

Signed-off-by: Jan Richter <jarichte@redhat.com>