Setting the test output_dir by spawners
Right now, the test output directory is managed by runners, and it is
separated from the `job/test-results` folder. Most of the runners are
using the temporary directory for test output, and this information
might be lost. Only possible way how to store this information to the
avocado results is use the avocado messages or use the `avocado jobs
get-output-files` script. For some spawners like `ProcessSpawner` this
is an unnecessarily complicated process, because we have to save the
same data twice even the tests are run in the same environment as
avocado.

This will make spawners smart enough to sett the output directory for
runners, so for the avocado will be easier to manage those files. For
example, the `ProcessSpawner` will set output dir to the
`job/test-reulst` so there is no need for creating the temporary
directory and copping to the `job/test-results`

Reference: #5083
Signed-off-by: Jan Richter <jarichte@redhat.com>