Instrumented tests with spawners output dir
It adds the usage of output directory created by spawners to the
instrumented runner. This also disables the legacy logging for the
nrunner which is managed by `Test` class. Those logs were hidden from
users, because they were created in temporal directory and didn't get
into the avocado `job/test-results` But now when we are not creating the
temp dir, those legacy logs are mixed up with the nrunners logs. That is
why we have to disable them.

Signed-off-by: Jan Richter <jarichte@redhat.com>