.travis.yml: save some build credits
Do not run static checks in travis, they're run in github actions
and this is enough since the statics checks are independant of the arch.

Not using 'make check' has also the advantage of not running
the targets clean and uninstall. This should save some build time
as well.

Do not set AVOCADO_PARALLEL_LINT_JOBS, it was used by the static
jobs.

Finally, none of the packages that were installed are required
anymore.

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>