drop output-check-record feature
This is working only on legacy runner that we don't support anymore.
Also, this it was deprecated at
a922b2421e69cb3c9d6e7ee043f73ea3265e1001. Since we are working on a new
feature with a complete new design based on the new runner architecture,
lets drop to avoid creating the new feature with legacy code.

Related to #3882.

Signed-off-by: Beraldo Leal <bleal@redhat.com>