pylint: fixes for W0237 (arguments-renamed)
Rename the variables and remove obsolete overrides when possible.
In the case of the class Test, add a decorator that allows
to continue using both parameters 'msg' and 'message', and
shows a deprecation warning when 'message' is used.

Make method 'fail' static to be consistent with the other
two methods 'error' and 'cancel'.

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>