docs/contributing: update info on parallel execution of tests
The AVOCADO_PARALLEL_CHECK has long been removed, and the parallel
execution of tests is now built into the nrunner implementation and
the appropriate selftest's suites.

With that, it doesn't make sense to suggest the execution of possibly
failing cases with Python's unittest runner (some tests may be Avocado
tests and should not be failing in parallel).

Signed-off-by: Cleber Rosa <crosa@redhat.com>