instrumented_tests: add support to Coverage
nrunner spawn sub-processes to run tests which is not good for Coverage
usage. Currently coverage is not working on nrunner. With this patch, we
bring it back. We detect if we are in coverage mode and using the
Coverage API we start, stop and save the measurements if running in
Coverage mode.

Fixes #4765.

Signed-off-by: Beraldo Leal <bleal@redhat.com>