Topological order of runtime tasks
This commit introduces the dependency graph. It connects RuntimeTasks by
its dependencies and sorts them by topological order in the
DependencyGraph.

The dependency graph ensures that test requirements won't collide,
because tests and requirements are represented by nodes in the graph and
the edges are representing dependence of test on requirement. Right now
we have only test and requirements, but the dependency graph supports
adding new categories of tasks like pre/post test task.

This is a part of BP004:
https://avocado-framework.readthedocs.io/en/latest/blueprints/BP004.html

Reference: #5203
Signed-off-by: Jan Richter <jarichte@redhat.com>