selftests/unit/test_tags.py: stress more of the tags values
This simply adds another test containing a tag with a value so that it
stresses out a little bit more the filtering code.

Also, and more importantly, it prepares the test and reduces noise
when a new filtering feature is added (being able to filter test that
do *not* have a given tag value).

Signed-off-by: Cleber Rosa <crosa@redhat.com>