requirement resolver tests enabled
The b42f35c7aed488c191f16cb57b2bfc20de410788 is solving the problems of
requirements tests from  #4620, so we can enable them. The only issue is
those are functional tests which can't be run in parallel because the
multiple executions of avocado with requirement runner is  breaking
avocado-software-manager. This commit solves this problem by running
them in serial. It splits the functional selftests into serial and
parallel and runs them in separated suite.

Reference: #4620
Signed-off-by: Jan Richter <jarichte@redhat.com>