lint.sh: always run all checks
selftests/lint.sh contains 3 different checks, if the first
one fails, the other are not run. This can lead to confusion
and to hide issues until the issues in the first command
are fixed.
With this change, all the pylint checks are run and the scripts
will fail if one of them failed.

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>