Plugins priority
Right now, the plugins are executed in alphabetical order. Such solution
might be problematic when the order of plugin executions is important
for plugin right behavior. This commit introduces `priority` variable
for plugins, which helps plugin developers to affect the order of plugin
executions. Now the plugins are sorted base on the priority, with higher
priority plugins are executed sooner and with lower priority plugins are
executed later.

Reference: #4939
Signed-off-by: Jan Richter <jarichte@redhat.com>