SPEC: better control and fix conditions of resultsdb package build
The current mechanism employed to control the build of the resultsdb
plugin are repeated across the file, but are now unified in a single
variable.

Also the conditions itself now accurately match the availability of
the python3-resultsdb_api in the respective distributions.

Reference: https://bodhi.fedoraproject.org/updates/FEDORA-EPEL-2021-1789f12fb5
Signed-off-by: Cleber Rosa <crosa@redhat.com>