nrunner: change BaseRunner interface
The interface enforced by nrunner.BaseRunner is one where the run()
method takes no arguments, and relies on the instance initialization
to supply a runnable.

This is incoherent when compared with other Avocado plugin interfaces,
where the actual required interfaces describe the data they need to
act on.

Also, all the information that a runner needs is in the Runnable, so
it should be possible in some cases (pending further work) to only
initialize on instance of a runner, and feed it a Runnable multiple
times.  For instance, the current registries of known *classes* of
runners, could actually be registries of instances.

In anyway, this will allow the nrunner.BaseRunner to be transformed
into a proper plugin interface, just like any other Avocado plugin.

Signed-off-by: Cleber Rosa <crosa@redhat.com>