nrunner: add interface tests for "task-run-recipe"
Recipe files allow the details of a task (in this case) to be
persisted and reused.  The "task-run-recipe" allows the execution
of those tasks from those files.

A task with a runnable of "golang" kind was not attempted to be run,
as it requires previous setup, such as environment variables or "go
env" commands that change the system.

Signed-off-by: Cleber Rosa <crosa@redhat.com>