nrunner: add interface tests for "runnable-run-recipe"
Recipe files allow the details of a runnable (in this case) to be
persisted and reused.  The "runnable-run-recipe" allows the execution
of those runnables from those files.

While many runners can handle different kinds of runnables, this
attempts to match specifc types based on the runner given. The
exception is when the runner does not match "avocado-runner-$type".

Also, the "golang" kind requires previous setup, such as environment
variables or "go env" commands that change the system, so it was
skipped here.

Signed-off-by: Cleber Rosa <crosa@redhat.com>