Nrunner output message
This introduces the new type of message for runners, the output message.
This message brings the ability for runners to print information to the
avocado console. The runner can send the output message and the avocado
will print its information to the console.

This is a backport of e2652b71668068838ea856bb7e16b262d34a511e.

Signed-off-by: Jan Richter <jarichte@redhat.com>