Settings: allow the command line help message to be more specific
In one specific situation solved here, the command line option is
much more limited in scope than the actual setting.

To avoid that disconnection, between the usage of the command line
options, and the underlying option, let's allow the command line
parser's help message to be given explicitly, if necessary.

Before this change, we'd have command line options following the
setting, such as:

 --disable-sysinfo  Enable or disable sysinfo information. Like hardware
                    details, profiles, etc.

Which is not accurate to the command line (there's no way to use
--disable-sysinfo to enable the feature), but is accurate to
underlying setting itself.  Now we can have a better description such
as:

 --disable-sysinfo  Disable sysinfo collection (like hardware details,
                    profiles, etc.)

Without impacting the configuration system and its automatically
generated documentation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>