Runners: drop "requirement" prefix from asset and package
These runners are already halfway destituted of their special
"requirement" prefix.  For instance, the runner scripts are already
named "avocado-runner-asset" and "avocado-runner-package".

TBH, while it makes it clear their intended use, there should be no
difference in their interface and implementation, and the fact that
declared requirements are fulfilled by them, is just an implementation
detail.

Signed-off-by: Cleber Rosa <crosa@redhat.com>