Asset/Package runner tests: use and equality check for the result
A presence check is not the ideal type of operation for the "result"
field, which must have an exact value.

While at it, reorganize the order of the arguments, because, IMO it
makes it easier to read along the lines of "assert that the result
obtained is equal to pass".

Signed-off-by: Cleber Rosa <crosa@redhat.com>