Legacy runner: remove the standalone MockingTest definition
The MockingTest is now left to support only a single user: the
TimeOutSkipTest.

Because of the legacy runner deprecation and soon removal, we can
simply merge MockingTest into TimeOutSkipTest and move it to the
plugin file.  That way, when the legacy plugin is removed,
TimeOutSkipTest which is now just and implementation detail, is also
removed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>