Process Spawner: forward PYTHONPATH to underlying process
When the runner command is "python3 -m avocado.core.runners.$type",
common for instance when running from Python eggs, that process needs
to know where Avocado (usually from an egg) is.

This is already done at the "check suitable runner stage", in
avocado/core/nrunner/runnable.py, but was lacking from the actual
spawning of the task.

Signed-off-by: Cleber Rosa <crosa@redhat.com>