nrunner: properly handled unsupported results
nrunner currently will simply trust the "result" that a runner will
produce in its "finished" message.  But this has side effects, and
should not proceed silently.

The legacy runner behavior, for instance, is to change it unknown
statuses to "ERROR", which is appropriate IMO.  But, it would also add
a "fail_reason" to the test information.

The difference proposed here, is to handle the details of such
violation internally, as an error message at the status repo level.
The user will only see that there was an error, but can then inspect
the logs to see why that was the case.  The justification is that this
will usually be caused by broken runners, and not broken tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>