avocado/core/loader.py: combine test loader classes
The SimpleFileLoader and FileLoader classes do not have a clear
separation.  For instance, the SIMPLE test type has always been
defined at the FileLoader class (when one would expect it to be in
SimpleFileLoader).

This reduces a number of duplicated methods, and ends up with a much
more sane implementation for the further removal of legacy code.

Signed-off-by: Cleber Rosa <crosa@redhat.com>