nrunner: properly handled unsupported results
nrunner currently will simply trust the "result" that a runner will
produce in its "finished" message.  But this has side effects, and
should not proceed silently.

The legacy runner behavior, for instance, is to change it unknown
statuses to "ERROR", which is appropriate IMO.  But, it  also adds
a "fail_reason" to the test information.

This implements the same behavior for nrunner.

Signed-off-by: Cleber Rosa <crosa@redhat.com>