runner: add JSON schema checking
This is an initial proposal for extending the interface checks, as far
as runners content/messages, from ad-hoc checks to a thorough schema
check.

This can and will be extended to check for the content of messages
produced at the various interfaces, such as "runnable-run",
"task-run", etc.

Reference: https://github.com/avocado-framework/avocado/issues/4058
Signed-off-by: Cleber Rosa <crosa@redhat.com>