selftests/functional/test_basic.py: drop legacy specific tests
The legacy runner starts a test process directly and waits for its
result.  This test checks for that specific behavior.

nrunner will use spawners, and its "wait_task()" will always be
involved in making sure the spawner finishes a task.

The current nrunner behavior, because of much looser coupling between
the task and a job, is to log that a task finished, while the result
related code will reconcile the missing test information as a CANCEL.

A revised behavior should be implemented for nrunner, and new tests
written.

Reference: https://github.com/avocado-framework/avocado/issues/5351
Signed-off-by: Cleber Rosa <crosa@redhat.com>