Removal of legacy runner
This removes the "runner" plugin, which is the main entry point into
a lot of legacy code.

It's not expected that every single bit of legacy code is removed
here, but I reckon it's a significant milestone to record a stage.

The usage of the loader in the suite code has also been removed, but
the loader code itself has not been removed yet.  The reason is that
the loader (avocado/core/loader.py) needs to have its
tests (selftests/unit/test_loader.py) either ported to the resolver or
confirmed that similar coverage exists.

Reference: https://github.com/avocado-framework/avocado/issues/5138
Signed-off-by: Cleber Rosa <crosa@redhat.com>