suite name contains slash fix
When the suite name contains a slash, the avocado crashes because the
`jobdata` module will try to save information to a folder with path
created from the suite name, which doesn't exist. This commit will use
`astring.string_to_safe_path()` for the jobdata folder path, which fixes
the issue.

Reference: #5365
Signed-off-by: Jan Richter <jarichte@redhat.com>