Discarding older messages from status repo fix
The status repo was discarding messages when the message timestamp was
lower than the latest save message. This change removes timestamp
checking and ensures the status changes only in expected life-cycle of a
test: started->running->finished. This means that change from running to
finished is possible, but from finished to running is not.

This change also stores every message into the `status_journal_summary`
list and sorts them by the timestamp. This sort is not optimal because,
the `runner_nrunner` plugin reads data from the status journal during
runner to server communication. This means that status repo sorts
messages which haven't been read yet by `runner_nrunner` plugin, but
can't ensure the order of the whole communication.

Reference: #5327
Signed-off-by: Jan Richter <jarichte@redhat.com>