Release workflow: work around git 2.34.4 changes
Similar to what was fixed in 56b1e653, git won't operate on a repo on
a directory owned by a different user, unless it's marked as safe.

The difference in this case, is that the error message is very
misleading:

   fatal: --local can only be used inside a git repository

A better error message should be produced by git, like the one given
when "git log" is called:

   fatal: unsafe repository (/path/to/repo is owned by someone else)

But this is something to be addressed on git itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>