task.timeout.running fix
In the current avocado version, the `task.timeout.running` option wasn't
working. Because we didn't have a mechanism for interrupting tasks after
the timeout reached. This commit adds an ability to spawners to
terminate the pawned task when it is necessary.

It fixes the #5383, but it also helps with solving problems with post
test plugins.

Reference: #5383
Signed-off-by: Jan Richter <jarichte@redhat.com>